openmemory-mcp/api/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Mem0 and ML dependencies
mem0ai==0.1.16
openai==1.3.7
qdrant-client==1.7.0

# Database and storage
sqlite-utils==3.35.1
sqlalchemy==2.0.23

# Utility dependencies
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0